+++ src/types/youtube/Next.ts
@@ -1,6 +1,9 @@
 import { YT_ItemSectionRenderer, YT_VideoPrimaryInfoRenderer, YT_VideoSecondaryInfoRenderer } from './Renderers';
 
 export type YT_NextApiResponse = {
+    responseContext: {
+        visitorData: string;
+    };
     contents: {
         twoColumnWatchNextResults: {
             results: {